Undump strings before yielding
The `ScriptExtractor`'s `extract_interpolated_values` method was
originally implemented by "borrowing" some code from the HAML parser
without taking the time to fully understand how it works.

It turns out that `String#dump` takes a string and escapes all of its
characters, so "something" becomes \"something\".

When the helper yielded an interpolated string, it was yielding the
dumped version, which ended up generating broken code. A hacky but
simple solution is to simply wrap the string in quotes and `eval` it to
convert it back to an unescaped string.

It's nasty, but it works.

Change-Id: I03fa843a6ee0dcad8e59b1c64eee40f9b130bc82
Reviewed-on: https://gerrit.causes.com/28452
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>