Add more Cops to the ignore list
While testing `haml-lint` on some production code with no `.rubocop.yml`
configuration file present, there were a lot of lints being reported
that were of little to no value.

`IfUnlessModifier` and `WhileUntilModifier` report lints for
if/unless/while/until blocks that only have one line. Since in HAML
it is possible to have one or no lines of actual Ruby in these
blocks, this lint will be reported for cases that aren't really
useful, like in the case of:

    - if test?
      %tag

Ignoring this cop removes any noise received from this lint.

`TrailingWhitespace` reports on trailing whitespace, which will also
happen with interpolation, for example:

    %p Some #{ interpolated_text } text

It'll be better to detect this ourselves rather than have Rubocop do
it since the lint message "trailing whitespace" isn't descriptive.

`Void` reports lints for lines of code that don't do anything, for
example having a line with just the string (not in any method or block):

    "Hello"

While usually valid, the `ScriptExtractor` abuses this in order to avoid
'assigned by unused variable' lints, which are arguably of higher value,
so we explicitly ignore this lint to eliminate noise.

Change-Id: Idc9607d29a62d39eafcdfbf7b306a2de652053fa
Reviewed-on: https://gerrit.causes.com/28453
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>