Fix indentation bug in ScriptExtractor
The `ScriptExtractor` had a bug where any `:ruby` filter containing a
block keyword (if, else, etc.) would have incorrect indentation
(potentially crashing if the filter was at the root level of the HAML
document).

The problem was that we were always subtracting one from the indentation
level when the line of code added contained a block keyword. Checking
that we only do this for actual tag nodes (instead of also arbitrary
lines of code from filters) fixes the problem.

Change-Id: I0d991a7b4c0d98c9d4fd8373af1a427610b2e063
Reviewed-on: https://gerrit.causes.com/30953
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>