Fix SpaceBeforeScript linter with nested tags
There was a bug with the `SpaceBeforeScript` linter where it would
incorrectly report a lint for a line with a space after the `=`/`-` if a
subsequent line contained the same substring but with no preceding
space.

For example:

  .div
    .another= key
  .again= something_with_the_word_key_in_it

This was due to a poorly thought-out algorithm for finding the "next
node" of a node. The faulty implementation was not recursive, and so
would not work for nested tags with inline script. The solution was to
implement a recursive `next_node` helper to solve the problem.

Fixes #1

Change-Id: I8acb47b5341cd299107fd8b1aadec80d3cef8ef2
Reviewed-on: https://gerrit.causes.com/32323
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>