Fix source mapping for multi-line interpolation
If the code extracted from interpolation in a filter spanned multiple
lines, the resulting source map would be incorrect due to it not taking
into consideration the additional lines (the code assumed only one line
could be added at a time).

Fix this by counting the number of newlines in the code and updating the
source map for each additional line to point to the original line that
generated it.

This has a problem of not pointing to the original line in the HAML
file, e.g. for the following code:

    :filter
      Something
      #{interpolation}

...any lints reported in the interpolated code will be mapped to the
containing `:filter`. To fix this comprehensively requires a fair amount
of work and tests to ensure correctness, which I'm not interested in
pursuing at this time.

Change-Id: I5037b6755fcdaae7b390641e91d8bf40fae8c191
Reviewed-on: http://gerrit.causes.com/38006
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>