Hack specs to work
I'm not proud of this, but my time is limited and unfortunately some
earlier decisions I've made, along with the change in direction of this
project to support configuration has made this code difficult to work
with.

Considering the simple high-level task this tool attempts to accomplish,
I'm introducing some tech debt in the interest of getting things
working.

Change-Id: I3fd05407f343a551729f8521c1fcefc2719bcdc2
Reviewed-on: http://gerrit.causes.com/40429
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>