Remove colorize dependency
We've had a few complaints about `haml-lint` pinning the version of
`colorize` to an old version (understandably so). This was done to fix a
change in behavior where `colorize` would still output color escape
sequences even when the output was redirected to a non-TTY.

At the cost of implementing colors ourself, this is now handled entirely
by the `Logger` class, so the dependency on `colorize` can be removed.

Change-Id: Iab226c4900266fdc12de4266eadfa9568cedee2c
Reviewed-on: http://gerrit.causes.com/40430
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>