Remove trailing HAML comment from parse tree
The HAML parser has a strange behavior where it inserts an empty HAML
comment at the end of a HAML document. This makes it weird when linters
are inspecting children of the root node, as there will always be an
extra child in the list.

Commit a08b381 originally attempted to deal with this by only addressing
it within the `HamlVisitor`. Unfortunately this was not sufficient as
linters can also inspect children directly via the `ParseNode`'s
`children` method.

Thus we resort to modifying the list of children for the actual root
node to exclude the trailing empty HAML comment. This actually is more
efficient since we aren't constantly checking if we're the root node
before deciding to exclude it. It's also cleaner since the modification
is managed entirely by the `Parser` class.

Change-Id: Ibab4463dba6cde17214b72d4db0dae77042c7671
Reviewed-on: http://gerrit.causes.com/42351
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>