Fix UnnecessaryStringOutput handling of nested content
The linter would incorrectly report lints for tags with nested content
containing interpolation, e.g.

  %tag
    Some #{interpolated} text

This is because the HAML parser treats the nested content as a script
node, even though it's not explicitly set as a script.

For now, include a special case in the linter. If more situations like
this come up, we'll add a special case to the `NodeTransformer` class.

Change-Id: I4000489c56d6e9b2f655a3c04154b43315bb64a2
Reviewed-on: http://gerrit.causes.com/42565
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>