Add TrailingWhitespace linter
Add linter which checks for trailing whitespace.

In order to get this to work, we had to alter the tests to not
automatically strip trailing whitespace from the `let(:haml)` blocks.
This resulted in having to adjust a few tests and also adding the
`AlignParameters` cop to the `RuboCop` ignore list (since the generated
code no longer removes inner whitespace).

Change-Id: If84c607532a45a2d341d08700dfb1fb36da65f51
Reviewed-on: http://gerrit.causes.com/44344
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>