Fix bundler warning about Rubocop dependency
When running `bundle install` inside this repository, a warning similar
to the following would be displayed:

  haml-lint at /Users/shane/src/haml-lint did not have a valid gemspec.
  This prevents bundler from installing bins or native extensions, but
  that may not affect its functionality.
  The validation message from Rubygems was:
    duplicate dependency on rubocop (= 0.27.1, development), (>= 0.25.0)
    use:
        add_runtime_dependency 'rubocop', '= 0.27.1', '>= 0.25.0'

This was because we were declaring the pinned version of RuboCop as a
development dependency in the gemspec. This turns out to not be
technically correct, as it's not really a development dependency, but a
testing one (i.e. we don't want to enforce this requirement for
developers using this gem when developing other gems that require this
gem).

It turns out a way get around this is to declare the pinned version in
the `Gemfile`, before the call to `gemspec` (you'll still get the
warning if you declare it after).

Change-Id: If6c312bbdc20b7eff3ca4242fc6a872bf35c0884
Reviewed-on: http://gerrit.causes.com/44721
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>