Fix LeadingCommentSpace linter failing on empty comment lines
If you comment out multiple lines in haml, there are often empty lines
which just have the '-#' for the comment. The TrailingWhitespace linter
does not allow a space at the end of the line but the
LeadingCommentSpace linter requires it.

This commit fixes the LeadingCommentSpace linter by allowing a comment
without a leading whitespace if the comment is empty.

Change-Id: Ibba637c7068d967b9c8b209ea90e18dad1d5c2a3
Reviewed-on: http://gerrit.causes.com/44884
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Joe Lencioni <joe.lencioni@brigade.com>