Explicitly set test target pattern to current file
In travis, bundled gems are located in the vendor folder in the current
directory. The default pattern scans every haml file recursively in the
current directory, so it finds offenses in bundled gems.

I fix this by setting the pattern to just the current file, which has (and
should always have) no offenses