Add JSONReporter
Using the `-r` or `--reporter` argument, now it's possible to specify a
reporter other than the default.  The `--show-reporters` argument
returns back the list of available reporters.  The Reporter class was
modified to statically keep a list of its descendants. See:
http://stackoverflow.com/a/7774881

For the detailed specification, see:
http://www.rubydoc.info/gems/rubocop/0.11.0#JSON_Formatter

`haml-lint` doesn't return the column where the offense is, so it's
omitted from the location hash.

Change-Id: Ib116914d7f00502503ae4e65bd0bf81cc33c99ee
Reviewed-on: http://gerrit.causes.com/46518
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Shane da Silva <shane.dasilva@brigade.com>