Report correct line numbers for interpolation in filters
The `RuboCop` linter was reporting incorrect line numbers for any cops
reported on code found in interpolation in filters, e.g.

    :filter
      This is some #{interpolation}
      Here's some #{more_interpolation}

We would report these cops as occurring on the `:filter` line, rather
than the line they appeared on. Fix this by determining the line number
during the extraction of interpolation code.

Fixes #75