Fix RuboCop linter to handle blocks with comments
When an anonymous block was trailed by a comment, it wouldn't be matched
as an anonymous block, and so a corresponding `end` statement would not
follow it, resulting in bad Ruby output. This ultimately lead to weird
errors reported by the `RuboCop` linter.

Fix it by expanding the regex to match `do` following by trailing
comments.