Support RuboCop 0.36
RuboCop previously depended on the astrolabe gem, but in
https://github.com/bbatsov/rubocop/commit/184d1163b5662e17383735f8d008b348dd6b9d34
the dependency was removed. It seems like this was a breaking change for
haml-lint, which continues to depend on astrolabe, but doesn't
explicitly install it as part of the haml-lint installation, so
haml-lint fails when it tries to require something which isn't there.