Add a linter to catch alignment tabs
Even when using tabs for indentation, it is generally accepted that
using tabs for alignment is something that should be avoided. Because
you can style your tabs to look differently in your editor, there is no
guarantee that your alignment tabs will show the code as aligned in
another person's editor.

This new linter detects when tabs are being used for alignment and warns
about the presence of the tab.

Closes #153