Add a linter for checking instance variable usage
It is fairly common to recommend that partials only use local variables
and helper methods and avoid the use of instance variables. This linter
checks views that follow partial conventions (prefixed with an
underscore) and checks that partials do not use instance variables.

The linter is also configurable, by default, to forbid the use of
instance variables in all views for those who wish to only use helper
methods in their views.

Lastly, it can be configured to accept any regular expression to match
against file names.

Closes #121