Add support for disabling linters with comments
When you are first introducing a linting system to a code base that
hasn't not previously used one, it can be a painful experience due to
the linter raising too many problems out of the gate. By enabling
linters to be selectively toggled off within a file, we can give our
users a way to gradually introduce linting into their code base.

This feature allows you do place a Haml comment within a file to toggle
a set of linters for a block. For example, adding this comment to the
top of a Haml file:

```haml
-# haml-lint:disable all
```

will disable all of the Haml-Lint linters for the remainder of the file,
unless later enabled with a similar comment in the file, like so:

```haml
-# haml-lint:enable AltText
```

The comment line can be either the string "all" or a comma-separated
list of linter names. They will be subsequently disabled or enabled
based on the comment. In the event of conflicting directives for
a linter, the last one wins.