Keep empty lines within Ruby filters
Because there are RuboCop rules that deal with spacing between blocks,
we need to retain empty lines within Ruby filters. Since Ruby filters
should be interpreted as pure Ruby, this does not change the semantics
of the output and allows us to use more of the cops without raising
false positives.