Report the real line with lint for LineLength
When we refactored the LineLength linter to be disableable, we missed
a corner case for multi-line nodes. These nodes were being reported at
line zero because the "find node for line" ability was ignoring the
presence of multiline nodes.

This fixes the ability to find lines within multiline nodes and ensures
the correct line is reported for the LineLength linter.

Closes #196