Set the current directory as default
When you don't specify any directory when calling `haml-lint`, it doesn't
inspect any file, while I would expect it to use the current directory to look
for files, as Rubocop does.

This is even more confusing when calling it with the option `auto-gen-config`,
as you have to call it like:

`haml-lint . --auto-gen-config`

which is really ugly, and it is not what the documentation says.