Fix finding attributes when assigned dynamically
In haml, you can assign attributes dynamically this way:

  %foo{ foo_attributes }

where `foo_attributes` is a method identifier or a variable. With the
newly added InlineStyles linter (619d581), we exposed a dormant bug with
how we look for attributes in a node in tag_node.rb. The following error
would be thrown when trying to lint the above code:

  undefined method `children' for false:HamlLint::ParsedRuby

I applied a somewhat naive fix by just returning early if `children`
isn't there.

Fixes #232