Fix handling of tags with mixed attribute styles
The `ClassAttributeWithStaticValue` linter would crash on tags that
mixed certain HTML- and hash-style attributes. This was caused by the
`dynamic_attributes_sources` method returning a string that started
with a `{` but was not a proper hash.

Fix by checking if the string also ends with a `}` so that we correctly
wrap the string in braces if it isn't already a hash.

Fixes #269