Fix InstanceVariables check for multiline blocks
Add surrounding statements and keywords to ruby script fragments
so parsed_script can return a valid abstract syntax tree.

Since parsed_script is only used by the InstanceVariable check it should
be sufficient to augment the script fragement with minimal dummy statements
to generate valid ruby syntax for the parser.  The returned AST may not
represent the actual document structure but it's suitable for detecting
usage of instance variables within the script fragment.

Append ';end' to script nodes with children to close their blocks.
As a special case we append ';when 0;end' to close case statements.
Prepend appropriate statements to script nodes with midblock keywords:
else, elsif, when, rescue, ensure