Enable RuboCop caching (#576)
* Remove unused Utils.with_captured_streams

* Enable RuboCop caching

Because we call RuboCop::Runner with the extracted ruby code on stdin, RuboCop thinks that its cache ought to be disabled.

Override the `RuboCop::Runner#cached_run?` method so that if haml-lint itself isn't using code from stdin, RuboCop is allowed to cache results based on a hash of the haml file.

Before:

```
$ bundle exec haml-lint app/{components,views}

626 files inspected, 0 lints detected

________________________________________________________
Executed in    5.06 secs    fish           external
   usr time   23.90 secs  340.00 micros   23.90 secs
   sys time    3.99 secs  993.00 micros    3.99 secs
```

After:

```
626 files inspected, 0 lints detected

________________________________________________________
Executed in    2.65 secs    fish           external
   usr time    8.78 secs    0.53 millis    8.78 secs
   sys time    2.19 secs    5.11 millis    2.18 secs
```

* Fix sporadic test failures from Tempfile cleanup